lab9/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
pyyaml>=6.0
dvc>=3.0.0
dvc-s3>=3.0.0
roar-cli

